AbstractShellProcessManager added, commonutils version up

+++ commonutils/src/main/java/net/maxsmr/commonutils/shell/RootShellCommands.java
@@ -1,22 +1,26 @@
 package net.maxsmr.commonutils.shell;
 
+import android.content.Context;
 import android.os.Build;
-import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
+import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
+import android.text.TextUtils;
 
 import net.maxsmr.commonutils.android.AppUtils;
-import net.maxsmr.commonutils.android.processmanager.AbstractProcessManager;
+import net.maxsmr.commonutils.android.processmanager.AbstractProcessManager;import net.maxsmr.commonutils.android.processmanager.model.ProcessInfo;
+import net.maxsmr.commonutils.data.FileHelper;
 import net.maxsmr.commonutils.logger.BaseLogger;
 import net.maxsmr.commonutils.logger.holder.BaseLoggerHolder;
 
+import org.jetbrains.annotations.NotNull;
+
 import java.io.File;
+import java.util.List;
+import java.util.Set;
 
 public final class RootShellCommands {
 
     private final static BaseLogger logger = BaseLoggerHolder.getInstance().getLogger(RootShellCommands.class);
-
-    private static final int MAX_INSTALL_RETRIES = 20;
-    private static final int MAX_UNINSTALL_RETRIES = 20;
 
     public RootShellCommands() {
         throw new AssertionError("no instances.");
@@ -25,12 +29,12 @@ public final class RootShellCommands {
     /**
      * Execute SU command
      */
-    @NotNull
+    @NonNull
     public static CommandResult executeCommand(String command) {
 
         try {
 
-            ShellWrapper rootShell = new ShellWrapper();
+            ShellWrapper rootShell = new ShellWrapper(false);
             CommandResult commandResult = rootShell.executeCommand(command, true);
             rootShell.dispose();
             return commandResult;
@@ -89,7 +93,7 @@ public final class RootShellCommands {
      * @param file File to read from
      * @return File's contents
      */
-    @Nullable
+    @android.support.annotation.Nullable
     public static String readFileAsRoot(File file) {
         String command = String.format("cat %s", file.getAbsolutePath());
 
@@ -122,26 +126,35 @@ public final class RootShellCommands {
      *
      * @param packageName If we need to update existing apk, you can pass the package name
      */
-    public static boolean installApk(File apkFile, @Nullable String packageName) {
-        String command;
+    public static boolean installApk(File apkFile, @android.support.annotation.Nullable String packageName, @NotNull Context context) {
 
+        if (!FileHelper.isFileCorrect(apkFile)) {
+            logger.e("Cannot install: incorrect apk file: " + apkFile);
+            return false;
+        }
+
+        String command;
         CommandResult commandResult;
 
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
             command = String.format("pm install -r -d %s", apkFile.getAbsolutePath());
         } else {
-            if (packageName != null) {
+            if (!TextUtils.isEmpty(packageName)
+                    && !packageName.equals(context.getPackageName())
+                    && AppUtils.isPackageInstalledFromShell(packageName)) {
                 command = String.format("pm uninstall -k %s", packageName);
                 commandResult = executeCommand(command);
-                logger.i("uninstall result: " + commandResult);
+                logger.i("Uninstall result: " + commandResult);
+                if (!commandResult.isSuccessful()) {
+                    return false;
+                }
             }
-
             command = String.format("pm install -r %s", apkFile.getAbsolutePath());
         }
 
         commandResult = executeCommand(command);
 
-        logger.i("install result: " + commandResult);
+        logger.i("Install result: " + commandResult);
 
         return commandResult.isSuccessful() && commandResult.getStdOut().toLowerCase().contains("success");
 
@@ -149,21 +162,34 @@ public final class RootShellCommands {
 
     /**
      * Try to install apk until success or max retries reached (workaround for some devices)
+     * @param maxRetriesCount 0 if install attempts should be infinite
      */
-    public static boolean installApkUntilSuccess(File apkFile, @Nullable String packageName) {
-        for (int i = 0; i < MAX_INSTALL_RETRIES; i++) {
-            if (installApk(apkFile, packageName)) {
-                return true;
-            }
+    public static boolean installApkUntilSuccess(int maxRetriesCount, File apkFile, @Nullable String packageName, @NotNull Context context) {
+        if (maxRetriesCount < 0) {
+            throw new IllegalArgumentException("Incorrect max retries count: " + maxRetriesCount);
         }
-
-        return false;
+        boolean result = false;
+        int tryCount = 0;
+        while (!result && maxRetriesCount == 0
+                || tryCount > maxRetriesCount) {
+            if (installApk(apkFile, packageName, context)) {
+                result = true;
+            }
+            tryCount++;
+        }
+        return result;
     }
 
     /**
      * Uninstall the given package via the Package manager
      */
     public static boolean uninstallPackage(String packageName) {
+
+        if (TextUtils.isEmpty(packageName)) {
+            logger.e("Cannot uninstall app: package name is empty");
+            return false;
+        }
+
         String command = "pm uninstall " + packageName;
 
         CommandResult commandResult = executeCommand(command);
@@ -180,29 +206,54 @@ public final class RootShellCommands {
 
     /**
      * Try to uninstall apk until success or max retries reached (workaround for some devices)
+     * @param maxRetriesCount 0 if uninstall attempts should be infinite
      */
-    public static boolean uninstallPackageUntilSuccess(String packageName) {
-        for (int i = 0; i < MAX_UNINSTALL_RETRIES; i++) {
-            if (uninstallPackage(packageName)) {
-                return true;
-            }
+    public static boolean uninstallPackageUntilSuccess(int maxRetriesCount, String packageName) {
+        if (maxRetriesCount < 0) {
+            throw new IllegalArgumentException("Incorrect max retries count: " + maxRetriesCount);
         }
-
-        return false;
+        boolean result = false;
+        int tryCount = 0;
+        while (!result && maxRetriesCount == 0
+                || tryCount > maxRetriesCount) {
+            if (uninstallPackage(packageName)) {
+                result = true;
+            }
+            tryCount++;
+        }
+        return result;
     }
 
     public static boolean killProcessByPid(int pid) {
         return RootShellCommands.executeCommand("kill -9 " + pid).isSuccessful();
     }
 
-    public static boolean killProcessByName(@Nullable String processName,
-                                            @NotNull AbstractProcessManager manager, boolean includeSystemPackages) {
-        final int pid = AppUtils.getPidByName(processName, manager, includeSystemPackages);
-        if (pid >= 0) {
-            return RootShellCommands.killProcessByPid(pid);
+    public static boolean killProcessesByName(@Nullable String processName,
+                                              @NotNull AbstractProcessManager manager, boolean includeSystemPackages) {
+        boolean result = true;
+        final Set<Integer> pids = AppUtils.getPidsByName(processName, manager, includeSystemPackages);
+        for (Integer pid : pids) {
+            result &= RootShellCommands.killProcessByPid(pid);
         }
-        return false;
+        return result;
     }
 
+    /**
+     * @return false if at least one kill was failed
+     */
+    public static boolean killApps(@Nullable List<String> apps,
+                                   @NotNull AbstractProcessManager manager, boolean includeSystemPackages) {
+        boolean result = true;
+        if (apps != null && !apps.isEmpty()) {
+            for (ProcessInfo process : manager.getProcesses(includeSystemPackages)) {
+                if (apps.contains(process.packageName)) {
+                    if (!killProcessByPid(process.pid)) {
+                        result = false;
+                    }
+                }
+            }
+        }
+        return result;
+    }
 
 }
